refactor: update redux actions import
