+++ src/new_achitecture/modules/mainPage/presentation/components/HeaderComponent/Nav/Nav.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { IconButton } from "@material-ui/core";
 import MenuIcon from "@material-ui/icons/Menu";
-import { setSidebarAction } from "../../../../../trip_search/presentation/redux/reducers/actions/app-actions";
+import { setSidebar} from "../../../../../../general/redux/slices/appSettingsSlice";
 import { useDispatch } from "react-redux";
 import css from "./Nav.module.css";
 import {useStyles} from "../../../../../../general/MUI/useStyles";
@@ -15,7 +15,7 @@ const Nav = () => {
             {/* <LanguageSelector />
             <currenciesSelector /> */}
             <IconButton
-                onClick={() => dispatch(setSidebarAction(true))}
+                onClick={() => dispatch(setSidebar(true))}
                 edge="end"
                 color="inherit"
                 aria-label="menu"
