+++ src/new_achitecture/modules/transfer/presentation/passengerPage/PassengerPage.jsx
@@ -3,8 +3,8 @@ import { Box, Container, LinearProgress, Typography } from "@material-ui/core";
 import i18next from "i18next";
 import { useSelector } from "react-redux";
 import { useDispatch } from "react-redux";
-import { getNextTransfersAction, getTransfersAction } from "../../../trip_search/presentation/redux/reducers/actions/transfers-actions";
-import { getLoading, getTransfersData } from "../../../trip_search/presentation/redux/reducers/selectors";
+import { getNextTransfersAction, getTransfersAction } from "../../../../general/redux/redux-using functions/transfers-actions";
+import { getLoading, getTransfersData } from "../../../../general/redux/selectors";
 import Transfer from "../transferPage/Transfer";
 import filtersClasses from "../../../../modules/trip_search/lib/filterSearch/Filter/FilterComponent.module.css";
 import classes from "./PassengerPage.module.css";
